+++ internal/repos/schedule_repo.go
@@ -14,7 +14,7 @@ func NewScheduleRepo(db *gorm.DB) *ScheduleRepo {
 	return &ScheduleRepo{db: db}
 }
 
-func (r *ScheduleRepo) CreateSchedule(schedule *models.Schedule) (uint, error) {
+func (r *ScheduleRepo) CreateSchedule(schedule *models.Schedule) (uint64, error) {
 	err := r.db.Create(schedule).Error
 	if err != nil {
 		return 0, err
@@ -23,7 +23,7 @@ func (r *ScheduleRepo) CreateSchedule(schedule *models.Schedule) (uint, error) {
 	return schedule.ID, nil
 }
 
-func (r *ScheduleRepo) AidNameExists(aidName string, userID uint) (bool, error) {
+func (r *ScheduleRepo) AidNameExists(aidName string, userID uint64) (bool, error) {
 	var count int64
 	err := r.db.Model(&models.Schedule{}).
 		Where("aid_name = ? AND user_id = ?", aidName, userID).
@@ -34,8 +34,8 @@ func (r *ScheduleRepo) AidNameExists(aidName string, userID uint) (bool, error) 
 	return count > 0, nil
 }
 
-func (r *ScheduleRepo) FindByUserID(userID uint) ([]uint, error) {
-	var scheduleID []uint
+func (r *ScheduleRepo) FindByUserID(userID uint64) ([]uint64, error) {
+	var scheduleID []uint64
 	err := r.db.Model(&models.Schedule{}).Where("user_id = ?", userID).Pluck("id", &scheduleID).Error
 	if err != nil {
 		return nil, err
@@ -43,7 +43,7 @@ func (r *ScheduleRepo) FindByUserID(userID uint) ([]uint, error) {
 	return scheduleID, nil
 }
 
-func (r *ScheduleRepo) FindSchedule(userID, scheduleID uint) (*models.Schedule, error) {
+func (r *ScheduleRepo) FindSchedule(userID, scheduleID uint64) (*models.Schedule, error) {
 	var schedule models.Schedule
 	err := r.db.Where("user_id = ? AND id = ?", userID, scheduleID).First(&schedule).Error
 	if err != nil {
@@ -52,7 +52,7 @@ func (r *ScheduleRepo) FindSchedule(userID, scheduleID uint) (*models.Schedule, 
 	return &schedule, nil
 }
 
-func (r *ScheduleRepo) NextTakings(userID uint) ([]models.Schedule, error) {
+func (r *ScheduleRepo) NextTakings(userID uint64) ([]models.Schedule, error) {
 	var schedules []models.Schedule
 	err := r.db.Where("user_id = ?", userID).Find(&schedules).Error
 	if err != nil {
